64 bit

+++ Engine/Graphics/Image.cpp
@@ -136,7 +136,7 @@ void TextureFrameTable::LoadAnimationSequenceAndPalettes(int uFrameID) {
     return;
 }
 
-int TextureFrameTable::FindTextureByName(const char *Str2) {
+int64_t TextureFrameTable::FindTextureByName(const char *Str2) {
     String name = Str2;
     std::transform(name.begin(), name.end(), name.begin(), ::tolower);
 
@@ -146,7 +146,7 @@ int TextureFrameTable::FindTextureByName(const char *Str2) {
     return -1;
 }
 
-Texture *TextureFrameTable::GetFrameTexture(int uFrameID, signed int a3) {
+Texture *TextureFrameTable::GetFrameTexture(int64_t uFrameID, signed int a3) {
     int v3 = uFrameID;
     TextureFrame *v4 = this->pTextures;
     TextureFrame *v5 = &v4[uFrameID];
